Refactor: 메세징 예외 분리, 메세징 반환 헤더에 메세지 타입 포함
