+++ src/main/java/com/knud4/an/exception/websocket/IllegalMessagingException.java
@@ -1,4 +1,4 @@
-package com.knud4.an.exception;
+package com.knud4.an.exception.websocket;
 
 public class IllegalMessagingException extends RuntimeException{
     public IllegalMessagingException(String message) {
